+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 5.9:
-  * English Translation of Sir Muhammad Zafrullah Khan ra     
+  * English Translation of Sir Muhammad Zafrullah Khan ra
+  * Bugs fixed related to matching of recent translation layers
 
 ## 5.8:
   * English Translation of Muhammad Ali ra added